Determine consensus based on all connected peers.
Extract the consensus calculation logic from the listing of peers
logic and use full dataset for the calculations, not the truncated
peer lists that the listing function deals with.

IPeersModule list() function was merged into getPeers() function as
now that the consensus counting isn't part of the list() anymore,
the functionality of the two was identical.