Remove internal caching of network state
Previous behaviour of caching was due to previous async nature of
the underlying peer APIs, but since now all of that is synchronous
it makes little sense to do the caching. Especially considering that
the cached peer data can get incorrect while it is cached and break
expectations of the code calling this method.